VOs, agregates and entities with its attributes
