+++ capitole-visibility-application/build.gradle
@@ -10,8 +10,14 @@ repositories {
 }
 
 dependencies {
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
+    implementation project(":capitole-visibility-domain")
+
+    implementation 'io.projectreactor:reactor-core:3.5.4'
+    implementation 'org.mockito:mockito-core:5.2.0'
+
+    testImplementation 'junit:junit:4.13.2'
+    testImplementation 'org.mockito:mockito-core:5.2.0'
+    testImplementation 'io.projectreactor:reactor-test:3.5.4'
 }
 
 test {
