+++ capitole-visibility-application/src/main/java/com/capitole/visibility/entity/Size.java
@@ -1,11 +1,13 @@
-package com.capitole.visibility.agregate;
+package com.capitole.visibility.entity;
 
+import com.capitole.visibility.agregate.Stock;
 import com.capitole.visibility.vo.SizeBackSoon;
 import com.capitole.visibility.vo.SizeId;
 import com.capitole.visibility.vo.SizeSpecial;
 
 public record Size(SizeId id,
                    SizeBackSoon sizeBackSoon,
-                   SizeSpecial sizeSpecial) {
+                   SizeSpecial sizeSpecial,
+                   Stock stock) {
 
 }
