+++ capitole-visibility-application/src/main/java/com/capitole/visibility/vo/SizeSpecial.java
@@ -1,4 +1,4 @@
 package com.capitole.visibility.vo;
 
-public record SizeBackSoon(boolean value) {
+public record SizeSpecial(boolean value) {
 }
