+++ capitole-visibility-application/src/main/java/com/capitole/visibility/vo/StockQuantity.java
@@ -1,4 +1,4 @@
 package com.capitole.visibility.vo;
 
-public record SizeSpecial(boolean value) {
+public record StockQuantity(boolean value) {
 }
